+++ example/lib/main.dart
@@ -7,36 +7,44 @@ void main() {
                    value: 😀,
                    unicode: 1F600,
                    name: Grinning Face,
+                   appleName: Grinning Face,
+                   alsoKnownAs: [Happy Face, Smiley Face],
                    group: Smileys & Emotion,
                    subgroup: face-smiling,
+                   version: 8.0,
+                   status: fully-qualified,
                    shortcodes: [
-                            Shortcode(
-                            platform: Discord,
-                            value: grinning
-                          ),
-                            Shortcode(
-                            platform: Github,
-                            value: grinning
-                          ),
-                            Shortcode(
-                            platform: Slack,
-                            value: grinning
-                          ),
-                            Shortcode(
-                            platform: Default,
-                            value: grinning_face
-                          ),
-                            Shortcode(
-                            platform: CLDR,
-                            value: grinning_face
-                          )]
+                        Shortcode(
+                        platform: Default,
+                        values: [grinning_face]
+                      ),
+                        Shortcode(
+                        platform: CLDR,
+                        values: [grinning_face]
+                      ),
+                        Shortcode(
+                        platform: Discord,
+                        values: [grinning]
+                      ),
+                        Shortcode(
+                        platform: Github,
+                        values: [grinning]
+                      ),
+                        Shortcode(
+                        platform: Slack,
+                        values: [grinning]
+                      )]
                   ),
                   Emoji(
                    value: 😃,
                    unicode: 1F603,
-                   name: Grinning face with big eyes,
+                   name: Grinning Face with Big Eyes,
+                   appleName: Grinning Face with Big Eyes
+                   alsoKnownAs: [Grinning Face, Happy Face, Happy, Smiley Face],
                    group: Smileys & Emotion,
                    subgroup: face-smiling,
+                   version: 6.0,
+                   status: fully-qualified
                    shortcodes: [...]
                   )
                  ...
@@ -49,7 +57,13 @@ void main() {
   // print(unicodes); // [1F600, 1F603, 1F604, 1F601, 1F606, 1F605, 1F923, 1F602, 1F642, ...]
 
   final names = emojis.names;
-  // print(names); // [Grinning face, Grinning face with big eyes, ...]
+  // print(names); // [Grinning Face, Grinning Face with Big Eyes, ...]
+
+  final appleNames = emojis.appleNames;
+  // print(appleNames); // [Grinning Face, Grinning Face with Big Eyes, ...]
+
+  final alsoKnownAs = emojis.alsoKnownAs;
+  // print(alsoKnownAs); // [Happy Face, Smiley Face, Grinning Face, ...]
 
   final shortcodes = emojis.shortcodes;
   // print(shortcodes); // [:grinning_face:, :grinning_face_with_big_eyes:, ...]
@@ -66,13 +80,37 @@ void main() {
   final slackShortcodes = emojis.slackShortcodes;
   // print(slackShortcodes); // [:grinning:, :smiley:, :smile:, :grin:, :laughing:, ...]
 
-  final groups = Emojis().groups;
-  // print(groups); // [Group.smileysAndEmotion, Group.peopleAndBody, Group.component, ...]
-  // print(groups.values); // [Smileys & Emotion, People & Body, Component, Animals & Nature, ...]
+  final groups = emojis.groups;
+  // print(groups); // [Group.smileysAndEmotion, Group.smileysAndEmotion, ...]
+  // print(groups.values); // [Smileys & Emotion, Smileys & Emotion, ...]
 
-  final subgroups = Emojis().subgroups;
-  // print(subgroups); // [Subgroup.faceSmiling, Subgroup.faceAffection, Subgroup.faceTongue, ...]
-  // print(subgroups.values); // face-smiling, face-affection, face-tongue, face-hand, ...]
+  final subgroups = emojis.subgroups;
+  // print(subgroups); // [Subgroup.faceSmiling, Subgroup.faceSmiling, ...]
+  // print(subgroups.values); // [face-smiling, face-smiling, ...]
+
+  final versions = emojis.versions;
+  // print(versions); // [Version.v8_0, Version.v6_0, Version.v6_0, ...]
+  // print(versions.values); // [8.0, 6.0, 6.0, 6.0, 6.0, 6.0, 9.0, ...]
+
+  final statuses = emojis.statuses;
+  // print(statuses); // [Status.fullyQualified, Status.fullyQualified, ...]
+  // print(statuses.values); // [fully-qualified, fully-qualified, ...]
+
+  final availableVersions = Emojis().versions;
+  // print(availableVersions); // [Version.v6_0, Version.v7_0, Version.v8_0, Version.v9_0, ...]
+  // print(availableVersions.values); // [6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 12.1, 13.0, ...]
+
+  final availableStatuses = Emojis().statuses;
+  // print(availableStatuses); // [Status.fullyQualified, Status.minimallyQualified, ...]
+  // print(availableStatuses.values); // [fully-qualified, minimally-qualified, unqualified, ...]
+
+  final availableGroups = Emojis().groups;
+  // print(availableGroups); // [Group.smileysAndEmotion, Group.peopleAndBody, Group.component, ...]
+  // print(availableGroups.values); // [Smileys & Emotion, People & Body, Component, Animals & Nature, ...]
+
+  final availableSubgroups = Emojis().subgroups;
+  // print(availableSubgroups); // [Subgroup.faceSmiling, Subgroup.faceAffection, Subgroup.faceTongue, ...]
+  // print(availableSubgroups.values); // face-smiling, face-affection, face-tongue, face-hand, ...]
 
   final groupsWithSubgroups = Emojis().groupsWithSubgroups;
   // print(groupsWithSubgroups); // {Group.smileysAndEmotion: [Subgroup.faceSmiling, ...], ...}
@@ -86,7 +124,7 @@ void main() {
   final name = Emojis().getOne('😀').name;
   // print(name); // Grinning face
 
-  final nameOrNull = Emojis().getOneOrNull('some broken value')?.name;
+  final nameOrNull = Emojis().getOneOrNull('some unknown value')?.name;
   // print(nameOrNull); // null
 
   final smileysAndEmotion = Emojis().smileysAndEmotion;
@@ -95,8 +133,8 @@ void main() {
   final faceSmiling = Emojis().faceSmiling;
   // print(faceSmiling); // [Emoji(value: 😀, ...), Emoji(value: 😃, ...), ...]
 
-  final any = '😀text😀'.emojis.any;
-  // print(any); // true
+  final contains = '😀text😀'.emojis.contains;
+  // print(contains); // true
 
   final only = '😀text😀'.emojis.only;
   // print(only); // true
@@ -115,6 +153,14 @@ void main() {
 
   final extract = '👍️te👍🏻xt👍🏼te👍🏽xt👍🏾te👍🏿xt'.emojis.extract;
   // print(extract); // [👍, 👍🏻, 👍🏼, 👍🏽, 👍🏾, 👍🏿]
+  // print(extract.first); // 👍
+  // print(extract.second); // 👍🏻
+  // print(extract.third); // 👍🏼
+  // print(extract.fourth); // 👍🏽
+  // print(extract.fifth); // 👍🏾
+  // print(extract.sixth); // 👍🏿
+  // print(extract.penultimate); // 👍🏾
+  // print(extract.last); // 👍🏿
 
   final emoji = '😀text'.emojis.get;
   // print(emoji);
@@ -122,36 +168,40 @@ void main() {
                           value: 😀,
                           unicode: 1F600,
                           name: Grinning Face,
+                          appleName: Grinning Face,
+                          alsoKnownAs: [Happy Face, Smiley Face],
                           group: Smileys & Emotion,
                           subgroup: face-smiling,
+                          version: 8.0,
+                          status: fully-qualified,
                           shortcodes: [
                             Shortcode(
-                            platform: Discord,
-                            value: grinning
-                          ),
-                            Shortcode(
-                            platform: Github,
-                            value: grinning
-                          ),
-                            Shortcode(
-                            platform: Slack,
-                            value: grinning
-                          ),
-                            Shortcode(
                             platform: Default,
-                            value: grinning_face
+                            values: [grinning_face]
                           ),
                             Shortcode(
                             platform: CLDR,
-                            value: grinning_face
+                            values: [grinning_face]
+                          ),
+                            Shortcode(
+                            platform: Discord,
+                            values: [grinning]
+                          ),
+                            Shortcode(
+                            platform: Github,
+                            values: [grinning]
+                          ),
+                            Shortcode(
+                            platform: Slack,
+                            values: [grinning]
                           )]
                         )] */
 
-  final hasAny = '👍️text'.emojis.hasAny(['👍']);
-  // print(hasAny); // true
+  final anyOf = '👍️text'.emojis.anyOf(['👍']);
+  // print(anyOf); // true
 
-  final hasEach = '👍️text'.emojis.hasEach(['👍']);
-  // print(hasEach); // true
+  final everyOf = '👍️text'.emojis.everyOf(['👍']);
+  // print(everyOf); // true
 
   final splitMapJoin = 'text😀text'.emojis.splitMapJoin(
         onMatch: (_) => '_emoji_',
@@ -162,9 +212,14 @@ void main() {
   final replace = '😀text👍'.emojis.replace('_');
   // print(replace); // _️text_
 
-  final replaceEach =
-      '😀text👍🏻text😀'.emojis.replaceEach({'😀': 'ABC', '👍🏻': '123'});
-  // print(replaceEach); // ABCtext123textABC
+  final replaceWith =
+      '😀text👍🏻text😀'.emojis.replaceWith({'😀': 'ABC', '👍🏻': '123'});
+  // print(replaceWith); // ABCtext123textABC
+
+  final replaceWhere = '😀text👍🏻text😀'
+      .emojis
+      .replaceWhere((e) => e.value == '👍🏻' ? '_OK_' : null);
+  // print(replaceWhere); // 😀text_OK_text😀
 
   final toShortcodes = '😀text🤦🏾‍♀️'.emojis.toShortcodes();
   // print(toShortcodes); // :grinning_face:text:woman_facepalming_medium_dark_skin_tone: