+++ lib/emoji_extension.dart
@@ -2,14 +2,14 @@ library emoji_extension;
 
 import 'package:emoji_extension/src/emojis/emoji_parser.dart';
 
+export 'src/emojis/emoji.dart';
 export 'src/emojis/emojis.dart';
 export 'src/emojis/group.dart';
+export 'src/emojis/shortcode.dart';
+export 'src/emojis/status.dart';
 export 'src/emojis/subgroup.dart';
-export 'src/extensions/emoji_list_extensions.dart';
-export 'src/extensions/emojis_extensions.dart';
-export 'src/extensions/group_list_extensions.dart';
-export 'src/extensions/shortcode_list_extensions.dart';
-export 'src/extensions/subgroup_list_extensions.dart';
+export 'src/emojis/version.dart';
+export 'src/extensions/extensions.dart';
 
 /// Extension on [String] that adds emoji parsing capabilities.
 extension EmojiExtension on String {