+++ lib/src/emojis/emoji.dart
@@ -2,7 +2,9 @@ import 'package:collection/collection.dart';
 import 'package:emoji_extension/src/emojis/group.dart';
 import 'package:emoji_extension/src/emojis/platform.dart';
 import 'package:emoji_extension/src/emojis/shortcode.dart';
+import 'package:emoji_extension/src/emojis/status.dart';
 import 'package:emoji_extension/src/emojis/subgroup.dart';
+import 'package:emoji_extension/src/emojis/version.dart';
 import 'package:emoji_extension/src/extensions/extensions.dart';
 
 /// Represents an Emoji.
@@ -10,8 +12,12 @@ import 'package:emoji_extension/src/extensions/extensions.dart';
 /// @param value The value of the emoji as a String.
 /// @param unicode The unicode representation of the emoji as a String.
 /// @param name The name of the emoji as a String.
+/// @param appleName The Apple name of the emoji as a nullable String.
+/// @param alsoKnownAs The other names emoji is also known as a List of String.
 /// @param group The group of the emoji as a Group enum.
 /// @param subgroup The subgroup of the emoji as a Subgroup enum.
+/// @param version The unicode version in which an emoji was introduced as a Version enum.
+/// @param status The status of the emoji as a Status enum.
 /// @param shortcodes The shortcodes of the emoji as a List of Shortcode objects.
 
 class Emoji {
@@ -20,8 +26,12 @@ class Emoji {
     required this.value,
     required this.unicode,
     required this.name,
+    this.appleName,
+    required this.alsoKnownAs,
     required this.group,
     required this.subgroup,
+    required this.version,
+    required this.status,
     required this.shortcodes,
   });
 
@@ -34,11 +44,23 @@ class Emoji {
   /// The name of the emoji as a String.
   final String name;
 
+  /// The Apple name of the emoji as a nullable String.
+  final String? appleName;
+
+  /// The list of names the emoji is also known as a List of String.
+  final List<String> alsoKnownAs;
+
   /// The group of the emoji as a Group enum.
   final Group group;
 
   /// The subgroup of the emoji as a Subgroup enum.
   final Subgroup subgroup;
+
+  /// The unicode version in which an emoji was introduced as a Version enum.
+  final Version version;
+
+  /// The status of the emoji as a Status enum.
+  final Status status;
 
   /// The shortcodes of the emoji as a List of Shortcode objects.
   final List<Shortcode> shortcodes;
@@ -54,31 +76,67 @@ class Emoji {
   /// Returns the default shortcode of the emoji as a String.
   String get shortcode => shortcodes.wherePlatform(Platform.Default)!;
 
+  /// Returns a new [Emoji] object with updated properties.
+  Emoji copyWith({
+    String? value,
+    String? unicode,
+    String? name,
+    String? appleName,
+    List<String>? alsoKnownAs,
+    Group? group,
+    Subgroup? subgroup,
+    Version? version,
+    Status? status,
+    List<Shortcode>? shortcodes,
+  }) {
+    return Emoji(
+      value: value ?? this.value,
+      unicode: unicode ?? this.unicode,
+      name: name ?? this.name,
+      appleName: appleName ?? this.appleName,
+      alsoKnownAs: alsoKnownAs ?? this.alsoKnownAs,
+      group: group ?? this.group,
+      subgroup: subgroup ?? this.subgroup,
+      version: version ?? this.version,
+      status: status ?? this.status,
+      shortcodes: shortcodes ?? this.shortcodes,
+    );
+  }
+
   /// Determines whether the current emoji is equal to another object.
   @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        other is Emoji &&
-            runtimeType == other.runtimeType &&
-            value == other.value &&
-            unicode == other.unicode &&
-            name == other.name &&
-            group == other.group &&
-            subgroup == other.subgroup &&
-            const DeepCollectionEquality.unordered()
-                .equals(shortcodes, other.shortcodes);
-  }
+  bool operator ==(Object o) =>
+      identical(this, o) ||
+      (o.runtimeType == runtimeType &&
+          o is Emoji &&
+          (identical(o.value, value) || o.value == value) &&
+          (identical(o.unicode, unicode) || o.unicode == unicode) &&
+          (identical(o.name, name) || o.name == name) &&
+          (identical(o.appleName, appleName) || o.appleName == appleName) &&
+          const DeepCollectionEquality.unordered()
+              .equals(o.alsoKnownAs, alsoKnownAs) &&
+          (identical(o.group, group) || o.group == group) &&
+          (identical(o.subgroup, subgroup) || o.subgroup == subgroup) &&
+          (identical(o.version, version) || o.version == version) &&
+          (identical(o.status, status) || o.status == status) &&
+          const DeepCollectionEquality.unordered()
+              .equals(o.shortcodes, shortcodes));
 
   /// Returns the hash code for the current emoji.
   @override
-  int get hashCode {
-    return value.hashCode ^
-        unicode.hashCode ^
-        name.hashCode ^
-        group.hashCode ^
-        subgroup.hashCode ^
-        const DeepCollectionEquality().hash(shortcodes);
-  }
+  int get hashCode => Object.hash(
+        runtimeType,
+        value,
+        unicode,
+        name,
+        appleName,
+        const DeepCollectionEquality().hash(alsoKnownAs),
+        group,
+        subgroup,
+        version,
+        status,
+        const DeepCollectionEquality().hash(shortcodes),
+      );
 
   /// Returns a String representation of the current emoji.
   @override
@@ -87,8 +145,12 @@ class Emoji {
         '  value: $value,\n'
         '  unicode: $unicode,\n'
         '  name: $name,\n'
+        '  appleName: $appleName,\n'
+        '  alsoKnownAs: $alsoKnownAs,\n'
         '  group: ${group.value},\n'
         '  subgroup: ${subgroup.value},\n'
+        '  version: ${version.value},\n'
+        '  status: ${status.value},\n'
         '  shortcodes: $shortcodes\n'
         ')';
   }