+++ lib/src/emojis/emojis.dart
@@ -3,11 +3,12 @@ import 'package:emoji_extension/src/emojis/emoji.dart';
 import 'package:emoji_extension/src/emojis/group.dart';
 import 'package:emoji_extension/src/emojis/platform.dart';
 import 'package:emoji_extension/src/emojis/shortcode.dart';
+import 'package:emoji_extension/src/emojis/status.dart';
 import 'package:emoji_extension/src/emojis/subgroup.dart';
+import 'package:emoji_extension/src/emojis/version.dart';
 import 'package:emoji_extension/src/extensions/extensions.dart';
 
-part 'emojis_json.dart';
-part 'json_mapper.dart';
+part 'emojis_data.dart';
 
 /// A class that represents a collection of Emojis.
 /// This class provides various methods to retrieve information about the Emojis, including
@@ -16,15 +17,13 @@ part 'json_mapper.dart';
 class Emojis {
   /// Creates a new instance of the Emojis class.
   /// This constructor should not be called directly. Instead, use the static factory
-  Emojis._() {
-    _emojis = _emojisJson.map((json) => json.toEmoji()).toUnmodifiableList();
-  }
+  const Emojis._();
 
   /// Returns a singleton instance of the Emojis class.
   factory Emojis() => _instance;
 
   /// The singleton instance of the Emojis class.
-  static final Emojis _instance = Emojis._();
+  static const Emojis _instance = Emojis._();
 
   /// Returns an unmodifiable list of all the Emojis.
   List<Emoji> get get => _emojis;
@@ -37,6 +36,12 @@ class Emojis {
 
   /// Returns an unmodifiable list of all the Emoji subgroups.
   List<Subgroup> get subgroups => Subgroup.values;
+
+  /// Returns an unmodifiable list of all the unicode emoji versions.
+  List<Version> get versions => Version.values;
+
+  /// Returns an unmodifiable list of all the Emoji statuses.
+  List<Status> get statuses => Status.values;
 
   /// Returns a map of all the Emoji groups with their corresponding subgroups.
   Map<Group, List<Subgroup>> get groupsWithSubgroups {
@@ -52,9 +57,9 @@ class Emojis {
     });
   }
 
-  /// Returns a list of all the platform names for Emoji shortcodes.
+  /// Returns an unmodifiable list of all the platform names for Emoji shortcodes.
   List<String> get shortcodePlatforms {
-    return Platform.values.map((e) => e.name).toList();
+    return Platform.values.map((e) => e.name).toUnmodifiableList();
   }
 
   /// Returns the Emoji with the specified value, unicode, name, or shortcode.
@@ -66,6 +71,8 @@ class Emojis {
           e.value == param ||
           e.unicode == param ||
           e.name.toLowerCase() == param.toLowerCase() ||
+          e.appleName?.toLowerCase() == param.toLowerCase() ||
+          e.alsoKnownAs.any((s) => s.toLowerCase() == param.toLowerCase()) ||
           e.shortcodes
               .any((s) => s.values.any((v) => v == param.removeColons())),
       orElse: orElse,
@@ -80,6 +87,8 @@ class Emojis {
           e.value == param ||
           e.unicode == param ||
           e.name.toLowerCase() == param.toLowerCase() ||
+          e.appleName?.toLowerCase() == param.toLowerCase() ||
+          e.alsoKnownAs.any((s) => s.toLowerCase() == param.toLowerCase()) ||
           e.shortcodes
               .any((s) => s.values.any((v) => v == param.removeColons())),
     );
@@ -97,7 +106,15 @@ class Emojis {
     return _emojis.where((e) => e.subgroup == subgroup).toUnmodifiableList();
   }
 
-  /// A List of all Emoji objects loaded from the JSON emoji data file,
-  /// used for searching and filtering.
-  late final List<Emoji> _emojis;
+  /// Returns a list of all [Emoji] objects in the _emojis list that belong
+  /// to the specified [Version].
+  List<Emoji> byVersion(Version version) {
+    return _emojis.where((e) => e.version == version).toUnmodifiableList();
+  }
+
+  /// Returns a list of all [Emoji] objects in the _emojis list that belong
+  /// to the specified [Status].
+  List<Emoji> byStatus(Status status) {
+    return _emojis.where((e) => e.status == status).toUnmodifiableList();
+  }
 }